+++ src/test/java/nextstep/subway/unit/domain/sections/SectionsAddTest.java
@@ -1,22 +1,18 @@
-package nextstep.subway.unit.domain;
+package nextstep.subway.unit.domain.sections;
 
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Section;
 import nextstep.subway.domain.Sections;
 import nextstep.subway.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
-class SectionsTest {
+class SectionsAddTest {
     private Line 이호선;
     private Station 강남역;
     private Station 삼성역;
@@ -139,46 +135,5 @@ class SectionsTest {
         // When & Then
         assertThatThrownBy(() -> sections.add(newSection))
                 .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("모든 역을 조회한다. 상행종착역부터 하행종착역까지의 순서를 보장한다")
-    @Test
-    void getStations() {
-        // Given
-        Section 하행종착역_뒤_section = new Section(이호선, 잠실역, 삼성역, 6);
-
-        sections.add(section);
-        sections.add(하행종착역_뒤_section);
-
-        // When
-        List<Station> stations = sections.getStations();
-
-        // Then
-        assertAll(
-                () -> assertThat(stations).hasSize(3),
-                () -> assertThat(stations).containsExactly(강남역, 잠실역, 삼성역)
-        );
-    }
-
-    @Test
-    void remove_exception_한개의_구간만_존재시() {
-        // Given
-        sections.add(section);
-
-        // When & Then
-        assertThatThrownBy(() -> sections.remove(강남역))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    void remove_exception_기준할_역이_없을때() {
-        // Given
-        sections.add(section);
-
-        Station 부산역 = new Station("부산역");
-
-        // When & Then
-        assertThatThrownBy(() -> sections.remove(부산역))
-                .isInstanceOf(IllegalStateException.class);
     }
 }
